(PC-25817)[PRO] refactor: Remove Breadcrumb component.
