--- pro/src/components/Breadcrumb/Breadcrumb.module.scss
@@ -1,74 +1,0 @@
-@use "styles/mixins/_fonts.scss" as fonts;
-@use "styles/mixins/_rem.scss" as rem;
-@use "styles/variables/_colors.scss" as colors;
-
-.pc-breadcrumb {
-  color: colors.$grey-medium;
-  display: flex;
-
-  .bc-step {
-    @include fonts.button;
-
-    color: colors.$grey-medium;
-    display: flex;
-    align-items: center;
-
-    a {
-      display: inline-flex;
-      color: colors.$black;
-
-      &:focus-visible {
-        outline: rem.torem(1px) solid colors.$input-text-color;
-        outline-offset: rem.torem(4px);
-        border-radius: rem.torem(4px);
-      }
-    }
-
-    &.active,
-    &.active a {
-      color: colors.$primary;
-      position: relative;
-
-      &::after {
-        border-bottom: rem.torem(2px) solid colors.$primary;
-        content: "";
-        left: 0;
-        position: absolute;
-        top: rem.torem(24px);
-        width: rem.torem(56px);
-      }
-    }
-
-    .error-icon {
-      margin-left: rem.torem(8px);
-      color: colors.$primary;
-    }
-  }
-
-  &.bc-tab {
-    color: colors.$black;
-
-    .bc-step {
-      @include fonts.button;
-
-      margin-left: rem.torem(24px);
-
-      a:hover,
-      a:focus-visible {
-        color: colors.$primary;
-      }
-
-      &:first-child {
-        margin-left: 0;
-      }
-    }
-  }
-
-  &.bc-disabled {
-    .bc-step:nth-child(1) {
-      a {
-        color: colors.$grey-dark;
-      }
-    }
-  }
-}
