--- pro/src/components/Breadcrumb/index.ts
@@ -1,2 +1,0 @@
-export { default, BreadcrumbStyle } from './Breadcrumb'
-export type { Step, StepPattern } from './types'
